replace change password Modal with modal common component, fixed imports for some common components
