+++ src/components/common/index.js
@@ -6,6 +6,7 @@ import Checkbox from './checkbox/Checkbox';
 import Logout from './logout/Logout';
 import Preloader from './preloader/Preloader';
 import Textarea from "./textarea/Textarea";
+import Modal from "./modal/Modal";
 
 export {
     Button,
@@ -15,5 +16,6 @@ export {
     Checkbox,
     Logout,
     Preloader,
-    Textarea
+    Textarea,
+    Modal
 }
