+++ src/components/common/modal/Modal.jsx
@@ -14,7 +14,8 @@ const Modal = ({modalTitle, isModalOpen, toggleModalVisibility, children}) => {
     bodyOverflow(isModalOpen);
     return (
         <>
-            {isModalOpen ? (<div className="common-modal">
+            {isModalOpen ?
+                <div className="common-modal">
                 <div className="common-modal__cover"/>
                 <div className="common-modal__dialog">
                     <div className="common-modal__content">
@@ -24,7 +25,8 @@ const Modal = ({modalTitle, isModalOpen, toggleModalVisibility, children}) => {
                         {children}
                     </div>
                 </div>
-            </div>) : null}
+            </div>
+                : null}
         </>
     );
 };
