+++ src/components/forms/ChangePassForm/ChangePassForm.scss
@@ -1,0 +1,7 @@
+.change-pass-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20px;
+  width: 250px;
+}
