--- src/components/modals/ChangePassModal/ChangePassModal.scss
@@ -1,25 +1,0 @@
-.modal {
-  position: absolute;
-  background-color: #fff;
-  box-shadow: 1px 1px 15px 1px rgba(0, 0, 0, 0.5);
-  padding: 40px;
-  border-radius: 5px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: end;
-  z-index: 5;
-
-  &__background {
-    position: absolute;
-    width: 100%;
-    height: 100vh;
-    top: 0;
-    left: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 4;
-    background-color: rgba(48,48,48,0.75);
-  }
-}
