+++ src/components/news/Dashboard.jsx
@@ -1,11 +1,9 @@
-import React, { Component } from "react";
-import { authService } from "../../services/auth.service";
-import { newsService } from "../../services/news.service";
-import Button from "../common/button/Button";
-import { Preloader } from "../common/index";
-import "./Dashboard.scss";
+import React, {Component} from "react";
+import {authService} from "../../services/auth.service";
+import {newsService} from "../../services/news.service";
+import {Button, Preloader, Modal} from "../common/index";
+import './Dashboard.scss'
 import PostItem from "./PostItem";
-import Modal from "../common/modal/Modal";
 import DashboardForm from "../forms/DashboardForm/DashboardForm";
 
 export default class Dashboard extends Component {
