+++ src/components/news/Post.jsx
@@ -2,8 +2,7 @@ import React, {Component} from 'react';
 import {authService} from "../../services/auth.service";
 import {newsService} from "../../services/news.service";
 import PostItem from "./PostItem";
-import Preloader from "../common/preloader/Preloader";
-import Modal from "../common/modal/Modal";
+import {Preloader, Modal} from "../common/index";
 import DashboardForm from "../forms/DashboardForm/DashboardForm";
 
 class Post extends Component {
